Add fixture, regenerate
